# Use buffer recycler for TOML buffers