--- a/toml/src/main/java/com/fasterxml/jackson/dataformat/toml/Parser.java
+++ b/toml/src/main/java/com/fasterxml/jackson/dataformat/toml/Parser.java
@@ -1,5 +1,6 @@
 package com.fasterxml.jackson.dataformat.toml;
 
+import com.fasterxml.jackson.core.io.IOContext;
 import com.fasterxml.jackson.core.io.NumberInput;
 import com.fasterxml.jackson.core.util.VersionUtil;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -27,15 +28,28 @@ class Parser {
 
     private TomlToken next;
 
-    private Parser(JacksonTomlParseException.ErrorContext errorContext, int options, Reader reader) throws IOException {
+    private Parser(
+            IOContext ioContext,
+            JacksonTomlParseException.ErrorContext errorContext,
+            int options,
+            Reader reader
+    ) throws IOException {
         this.errorContext = errorContext;
         this.options = options;
-        this.lexer = new Lexer(reader, errorContext);
+        this.lexer = new Lexer(reader, ioContext, errorContext);
+        lexer.prohibitInternalBufferAllocate = (options & TomlWriteFeature.INTERNAL_PROHIBIT_INTERNAL_BUFFER_ALLOCATE) != 0;
         this.next = lexer.yylex();
     }
 
-    public static ObjectNode parse(JacksonTomlParseException.ErrorContext errorContext, int options, Reader reader) throws IOException {
-        return new Parser(errorContext, options, reader).parse();
+    public static ObjectNode parse(
+            IOContext ioContext,
+            int options,
+            Reader reader
+    ) throws IOException {
+        Parser parser = new Parser(ioContext, new JacksonTomlParseException.ErrorContext(ioContext, null), options, reader);
+        ObjectNode node = parser.parse();
+        parser.lexer.releaseBuffers();
+        return node;
     }
 
     private TomlToken peek() throws JacksonTomlParseException {