--- a/toml/src/main/java/com/fasterxml/jackson/dataformat/toml/TomlFactory.java
+++ b/toml/src/main/java/com/fasterxml/jackson/dataformat/toml/TomlFactory.java
@@ -308,13 +308,12 @@ public final class TomlFactory extends JsonFactory {
      */
 
     private ObjectNode parse(IOContext ctxt, Reader r0) throws IOException {
-        JacksonTomlParseException.ErrorContext errorContext = new JacksonTomlParseException.ErrorContext(ctxt.getSourceReference(), null);
         if (ctxt.isResourceManaged() || isEnabled(StreamReadFeature.AUTO_CLOSE_SOURCE)) {
             try (Reader r = r0) {
-                return Parser.parse(errorContext, _tomlParserFeatures, r);
+                return Parser.parse(ctxt, _tomlParserFeatures, r);
             }
         } else {
-            return Parser.parse(errorContext, _tomlParserFeatures, r0);
+            return Parser.parse(ctxt, _tomlParserFeatures, r0);
         }
     }
 }