This PR makes the jflex-generated Lexer use allocTokenBuffer for the main token buffer (ba4f8edd47b42daf913447b50d20ae90b1c77759) and constructTextBuffer for string building (c042f4e5f6943eaad1fba1145ecc31362c039d83).

This is accomplished by altering the default jflex skeleton ( https://github.com/jflex-de/jflex/blob/4edfabcaa7eb06a41f8429a7462eca6dc92d7247/jflex/src/main/resources/jflex/skeleton.default ) to remove the character array allocations, and implementing those using IOContext in the .jflex instead. Needing to copy the whole skeleton isn't ideal, but oh well.

This approach only works if individual jflex tokens are smaller than the default buffer returned by allocTokenBuffer. When there are very long tokens, we fall back on new char[]. We cannot use allocTokenBuffer(minSize), because we need to keep the old buffer available, and there cannot be two token buffers in use at the same time from the same IOContext.

Ideally, we would never need to resize the buffers at all. Some of the potential long tokens (maybe all?) are tested in LongTokenTest.

- long numbers could maybe just throw an exception. For json, jackson writes long numbers to the text buffer as fallback, but any "fallback" behavior that only happens with long inputs is difficult with jflex, because DFAs can't count. It'd require creating a new state that implements this behavior.
- comments and whitespace could be ignored, since we never need to access the content of those tokens. Could be solved with a new state fairly easily, if jflex cooperates.
- strings are already copied to the text buffer, and this logic could be improved. Right now, long sequences of unescaped data are processed as a single token for performance, but maybe there's an alternative here too. The "fallback" behavior problems for numbers apply.

All in all I'm not awfully concerned about this problem. None of these cases are likely to occur for "benign" data; even string and comment data is regularly "flushed" as long as there's a newline somewhere in between. So, the only issue here is that someone could build a malicious TOML file that takes a lot of memory to parse, but even then the attack is not very effective, because at most 4x as much memory is used as there is data (2x for utf-8 -> char array, 2x for the resize factor). There are definitely better attacks, e.g. by creating many json nodes (since we cache the whole tree during parsing).

For this reason, imo, it is not worth addressing this beyond what is already in this PR.