@yawkat 2021-04-10
jackson-benchmark results for the 2.13 version of this patch:

before:
```
Benchmark                                    Mode  Cnt       Score       Error  Units
TOMLStdReadAfterburner.readPojoMediaItem    thrpt    7  123674.553 ±  2572.701  ops/s
TOMLStdReadVanilla.readPojoMediaItem        thrpt    7  119795.915 ±  1100.263  ops/s
TOMLStdWriteAfterburner.writePojoMediaItem  thrpt    7  496205.741 ±  7262.247  ops/s
TOMLStdWriteVanilla.writePojoMediaItem      thrpt    7  469790.847 ± 11204.921  ops/s
```

after:
```
Benchmark                                    Mode  Cnt       Score      Error  Units
TOMLStdReadAfterburner.readPojoMediaItem    thrpt    7  134938.605 ± 4316.725  ops/s
TOMLStdReadVanilla.readPojoMediaItem        thrpt    7  129000.369 ±  994.874  ops/s
TOMLStdWriteAfterburner.writePojoMediaItem  thrpt    7  515883.030 ± 7075.571  ops/s
TOMLStdWriteVanilla.writePojoMediaItem      thrpt    7  475328.670 ± 2755.866  ops/s
```

Probably spent more time building this patch than would ever be saved by the improvement :D