@yawkat 2021-04-10
Oh, one thing that was a bit weird @cowtowncoder , I had to use `textBuffer.emptyAndGetCurrentSegment();` instead of just `resetWithEmpty`, even though I think both are supposed to be equivalent. resetWithEmpty sets the currentSegment to null, leading to an NPE in any subsequent append calls. Am I using it wrong or is that a bug?