@cowtowncoder 2021-04-10
@yawkat First of all, thank you for figuring this out! I was wondering if this was possible.

I also recognize the feeling of doing some optimization, only to find its effects are less than what profiler might have suggested. Still, +10% is not too shabby in general and sometimes there's nice stacking effect at some point (some combination of smaller changes manages to compound to something bigger).

I'll have a look at what's up with `TextBuffer`, there may be some odd corner case with initialization, some assumption code has on call patterns. `resetWithEmpty()` should usually be sufficient in cases where value of "empty String" is implied. But `emptyAndGetCurrentSegment()` is the expected call when starting to build new Strings.