@cowtowncoder 2021-04-10
Ok, actually, code as it is now is correct. `resetWithEmpty()` javadocs could probably use improvement as it really should only be used to indicate the case of "current text collected is empty String" -- but it is not reset to start where collection can begin. Essentially all `resetXxx()` methods mean "set buffer to contain value X".

But one thing I could consider here is ... optimizing `TextBuffer.append()`. Most existing usage just gets underlying `char[]` segment for direct modification, but that does not work with Lexer hear, as it's state machine based, and `append()` makes more sense. I suspect I could at least try changing things a bit to streamline the happy path and see if I can observe actual improvement.
Usually I might not bother but I suspect this is a hot spot now, on critical path. And code in question is small enough that I can play with it, run tests etc. If it doesn't matter should be easy to find out. :)