@cowtowncoder 2021-04-11
Oh. Should have looked at profiler output bit more... I don't think single-char `TextBuffer.append()` is actually used a lot.
Most usage via bulk appends. Will see if something can be done there, still. Maybe there's some room for optimization.

Another sort of interesting thing is that on reader side, as all input will become `ObjectNode` at least briefly, it might be possible to make `ObjectMapper.readTree()` avoid tree-from-tree construction.
The place to do this would probably be `JsonNodeDeserializer` which could detect use of `TreeTraversingParser` and then just use the copy? The only question is whether use of custom sub-class (of ObjectNode/ArrayNode) might be problematic or not.
Not 100% sure how much overhead this would avoid, potentially not a lot... but there is some overhead in traversal and recreation of the whole tree.